refactor(spring-keycloak-example): reorganize package structure for controllers and security configurations
