+++ spring-keycloak-example/src/main/java/com/io/example/oidcAuth/config/OIDCSecurityConfig.java
@@ -1,20 +1,26 @@
-package com.io.example.config;
+package com.io.example.oidcAuth.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 
-@Profile("default")
+@Profile("OIDC")
 @Configuration
 @SuppressWarnings("unused")
-public class DefaultSecurityConfig {
+public class OIDCSecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http.authorizeHttpRequests(auth -> auth.anyRequest().permitAll())
-                   .build();
-    }
+        http
+            .authorizeHttpRequests(auth -> auth
+                .requestMatchers("/public").permitAll()
+                .anyRequest().authenticated()
+            )
+            .oauth2Login(Customizer.withDefaults());
 
+        return http.build();
+    }
 }
