+++ spring-keycloak-example/src/main/java/com/io/example/standard/config/DefaultSecurityConfig.java
@@ -1,26 +1,20 @@
-package com.io.example.config;
+package com.io.example.standard.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.web.SecurityFilterChain;
 
-@Profile("OIDC")
+@Profile("default")
 @Configuration
 @SuppressWarnings("unused")
-public class OIDCSecurityConfig {
+public class DefaultSecurityConfig {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http
-            .authorizeHttpRequests(auth -> auth
-                .requestMatchers("/public").permitAll()
-                .anyRequest().authenticated()
-            )
-            .oauth2Login(Customizer.withDefaults());
-
-        return http.build();
+        return http.authorizeHttpRequests(auth -> auth.anyRequest().permitAll())
+                   .build();
     }
+
 }
