+++ spring-keycloak-example/src/main/java/com/io/example/standard/controller/DefaultHomeController.java
@@ -1,4 +1,4 @@
-package com.io.example.controller;
+package com.io.example.standard.controller;
 
 import org.springframework.context.annotation.Profile;
 import org.springframework.web.bind.annotation.GetMapping;
